Unsupervised Machine Learning
From the given ‘Iris’ dataset, we have to predict the optimum number of clusters and represent it visually. To find the optimum number of clusters we will use the elbow method to find the number of clusters and K-Means clustering.
The Iris Dataset contains four features (length and width of sepals and petals) of 150 samples of three species of Iris (Iris setosa, Iris virginica and Iris versicolor).